democracy-funding/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
pydantic>=2.9.0
pydantic-settings>=2.0.0
sqlalchemy>=2.0.0
aiosqlite>=0.20.0
httpx>=0.27.0
anthropic>=0.40.0
pyyaml>=6.0
bcrypt>=4.0.0
python-multipart>=0.0.6
aiofiles>=23.0.0
